Make names of new functions consistent
Rename process_mth5_dev to process_mth5_new
Deleted old process_mth5_new which was an abandoned module
Modified all imports appropraitely

[Issue(s): #153]
